fix: address downgrade cli review
