+++ website/app/docs/cli/page.mdx
@@ -935,13 +935,13 @@ For the HTTP endpoint version, see [MCP Server](/docs/customization/mcp).
 
 ## Supported Frameworks
 
-| Framework | Detection | `init` | `upgrade` |
-| --------- | --------- | ------ | --------- |
-| Next.js | `next` in dependencies | Supported | Supported |
-| TanStack Start | `@tanstack/react-start` in dependencies | Supported | Supported |
-| SvelteKit | `@sveltejs/kit` in dependencies | Supported | Supported |
-| Astro | `astro` in dependencies | Supported | Supported |
-| Nuxt | `nuxt` in dependencies | Supported | Supported |
+| Framework | Detection | `init` | `upgrade` | `downgrade` |
+| --------- | --------- | ------ | --------- | ----------- |
+| Next.js | `next` in dependencies | Supported | Supported | Supported |
+| TanStack Start | `@tanstack/react-start` in dependencies | Supported | Supported | Supported |
+| SvelteKit | `@sveltejs/kit` in dependencies | Supported | Supported | Supported |
+| Astro | `astro` in dependencies | Supported | Supported | Supported |
+| Nuxt | `nuxt` in dependencies | Supported | Supported | Supported |
 
 ## Available Themes
